pyproject.toml
[tool.poetry]
name = "credit-scoring"
version = "0.1.0"
description = "Alternative Credit Scoring Platform"
authors = ["befikadusata <befikadusata@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
pandas = ">=2.3.3,<3.0.0"
numpy = ">=2.3.5,<3.0.0"
matplotlib = ">=3.10.7,<4.0.0"
seaborn = ">=0.13.2,<0.14.0"
jupyter = ">=1.1.1,<2.0.0"
requests = ">=2.32.5,<3.0.0"
scikit-learn = ">=1.7.2,<2.0.0"
papermill = ">=2.6.0,<3.0.0"
joblib = ">=1.5.2,<2.0.0"
fastapi = ">=0.123.9,<0.124.0"
uvicorn = ">=0.38.0,<0.39.0"
prometheus-fastapi-instrumentator = ">=7.1.0,<8.0.0"
redis = ">=7.1.0,<8.0.0"
mlflow = ">=3.7.0,<4.0.0"
optuna = ">=4.6.0,<5.0.0"
xgboost = ">=3.1.2,<4.0.0"
python-json-logger = ">=4.0.0,<5.0.0"
json-logger = ">=0.0.1,<0.0.2"
psycopg2-binary = ">=2.9.11,<3.0.0"
boto3 = ">=1.42.4,<2.0.0"
apache-airflow = ">=3.1.0,<4.0.0"
apache-airflow-providers-docker = ">=3.0.0,<4.0.0"
evidently = ">=0.4.0,<1.0.0"
prometheus-client = ">=0.19.0,<1.0.0"
kafka-python = ">=2.0.2,<3.0.0"
shap = ">=0.44.0,<0.45.0"
pydantic = ">=2.10,<3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = ">=25.11.0,<26.0.0"
pytest = ">=9.0.1,<10.0.0"
ruff = ">=0.14.9,<1.0.0"
pre-commit = ">=4.5.0,<5.0.0"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "A"]
ignore = ["E501", "E402", "F841"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true
lines-after-imports = 2

[tool.black]
line-length = 88